Move uploader job classes for consistency.
