+++ mash/services/uploader/azure_job.py
