+++ mash/services/uploader/ec2_job.py
