+++ mash/services/uploader/gce_job.py
